requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
opencv-python>=4.9.0
scikit-image>=0.22.0
scikit-learn>=1.4.0
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.12.0
Pillow>=10.0.0
